Refactoring and test cleanup.
Did some refactoring to help make tests, and code in general, easier to read. This largely involves swapping out arrays for collections in a number of places. Fixed a few minor bugs along the way. Added some functionality to Spellbook and SpellbookUtils to help dealing with when Sources were added to the app.